Maven Settings
This change adds a BP_MAVEN_SETTINGS environment variable that allows a user
to pass in the contents of a Maven settings.xml file.  The contents of that
file are then written to the filesystem and used with the `--settings` flag
when invoking Maven.  The environment variable will allow users to specify a
settings.xml file without checking it into source control.

Signed-off-by: Ben Hale <bhale@pivotal.io>