Rebuild application if settings.xml changes
Resolves #3

The contents of a setting.xml file provided by a maven binding can impact
the result of a build. The cached jar should be invalidated if the contents of
settings.xml changes. Therefore, the maven buildpack now adds the hash of
settings.xml to the application layer metadata.

Updates libbs 1.1.2 -> 1.2.1

Signed-off-by: Emily Casey <ecasey@pivotal.io>