Configurable install & build support
Based on what is in the buildplan, the buildpack can now optionally install and optionally build.

If `maven` is in the buildplan, then the buildpack will attempt to install or validate that Maven is installed and generate a command to run Maven.

This comes in four forms:

1. Install Maven Daemon
2. Install Maven
3. Validate Maven Wrapper
4. Validate Maven on $PATH

If `jvm-application-package` is in the buildplan, then the buildpack will attempt to perform a build with Maven. This uses the command from the previously installed or validated Maven.

This enables four scenarios:

1. The buildpack installs & runs Maven (or Maven Daemon)
2. The buildpack runs Maven w/Maven Wrapper
2. The buildpack just installs Maven (or Maven Daemon)
3. The buildpack doesn't install Maven or use Maven Wrapper, but just runs Maven (Maven is provided through other means)

Signed-off-by: Daniel Mikusa <dan@mikusa.com>