Fix bug where command is not set
If Maven was not in the buildplan, then the command would never be set. This means the case where Maven is provided externally would never work.

The adjusted logic will unconditionally run the NoopMavenManger when maven is not in the buildplan. This will try to find Maven on the PATH. If it cannot find it on the path, then you get a reasonable error.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>