Adjust detect logic for non-Java apps
There was a bug with the detect logic where the Maven buildpack was providing 'jvm-application-package' without any requirements. This paired with executable-jar and the application server buildpacks which had a requirement of 'jvm-application-package' and created a successful buildplan when there shouldn't have been one. This was originally added to support the case where someone might want to perform a build but not have the buildpack install Maven.

This PR changes the detection logic so that the 'jvm-application' is only added if a pom.xml file exists. This limits the functionality a bit, as you can only do a build-only build if you have a pom.xml but that seems reasonable at this time.

If necessary, we could circle back and adjust the logic further, but the plan is to hold off until there is a more defined used case that would require this functionality.

Additionally, this PR converts ioutil usage which has been deprecated.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>