Add Maven 4
This commit adds Maven 4 (alpha). This is added as a second major branch for anyone that is interested in trying it.

It also adds the ability to select the major version of Maven that is used. The default version is set to version 3.x. By setting `BP_MAVEN_VERSION`, you can override this.

This also includes a minor fix where the logger was not being plumbed through to the Maven Manager object. That could have resulted in loss of some log events.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>