Forgot double quote around regex
I skipped the double quotes around the regex. This meant the escaped backslashes were not unescaped, which breaks the regex.

Regular expression patterns in yaml files should either be quoted, in which case you need to escape backslashes, or not quoted and then you do not need to escape the backslashes.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>